refactor(stack): move sidecar.Options() to stack package
